requirements.txt
flask
flask-sqlalchemy
flask-bootstrap
flask_httpauth
black
pytest
requests
python-dateutil